WooCommerce V3.x.x Compatible
In this commit we have fixed the warning displyed on the Print screens, on the admin side it was shwoing notice on the WC orders columns.

Also, when customer click on the print view from the email then the notices was displayed

Also, the My Account page was showing the notices on the Print

All these notices has been fixed.

It is also made compatible with the latest WC version as well as with Older version